refactor: extract SendFundsAmountForm sub-components into own files
